+++ app/services/immich/request_photos.rb
@@ -5,7 +5,7 @@ class Immich::RequestPhotos
 
   def initialize(user, start_date: '1970-01-01', end_date: nil)
     @user = user
-    @immich_api_base_url = "#{user.settings['immich_url']}/api/search/metadata"
+    @immich_api_base_url = URI.parse("#{user.settings['immich_url']}/api/search/metadata")
     @immich_api_key = user.settings['immich_api_key']
     @start_date = start_date
     @end_date = end_date