+++ app/services/imports/create.rb
@@ -24,12 +24,12 @@ class Imports::Create
   def parser(source)
     # Bad classes naming by the way, they are not parsers, they are point creators
     case source
-    when 'google_semantic_history'  then GoogleMaps::SemanticHistoryParser
-    when 'google_phone_takeout'     then GoogleMaps::PhoneTakeoutParser
-    when 'owntracks'                then OwnTracks::ExportParser
-    when 'gpx'                      then Gpx::TrackParser
-    when 'immich_api'               then Immich::ImportParser
-    when 'geojson'                  then Geojson::ImportParser
+    when 'google_semantic_history'      then GoogleMaps::SemanticHistoryParser
+    when 'google_phone_takeout'         then GoogleMaps::PhoneTakeoutParser
+    when 'owntracks'                    then OwnTracks::ExportParser
+    when 'gpx'                          then Gpx::TrackParser
+    when 'geojson'                      then Geojson::ImportParser
+    when 'immich_api', 'photoprism_api' then Photos::ImportParser
     end
   end